+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ## [Unreleased]
 
+## [3.2.0] - 2026-01-15
+
+### Added
+- Added new 'get_run_log' tool that allows to get logs from run configurations in both 'run' and 'debug' modes.
+
+### Changed
+- updated 'list_debug_sessions' tool now shows both 'run' and 'debug' mode sessions
+
 ## [3.1.0] - 2026-01-07
 
 ### Added